requirements.txt
# 標準ライブラリ以外の必要パッケージ
boto3
python-dotenv
numpy==1.26.4
scipy
opencv-python
# カルマンフィルタ用
filterpy
# GUIで出したい場合はシステム版opencv( sudo apt install python3-opencv )を使用
websockets
# WebRTC配信プロセス用
aiortc
aiohttp
av

# ローカルIP通知APIPost用
requests

# boto3の依存
botocore